refactor(eval): separate result processing from models
