+++ src/evaluator/inMemoryStore.ts
@@ -1,3 +1,5 @@
+import { getResultIndexKey } from './resultIndex';
+
 import type {
   CompletedPrompt,
   EvaluateResult,
@@ -21,10 +23,6 @@ export interface InMemoryEvaluation {
   durationMs?: number;
   generationDurationMs?: number;
   evaluationDurationMs?: number;
-}
-
-function getResultIndexKey(result: Pick<EvaluateResult, 'promptIdx' | 'testIdx'>): string {
-  return `${result.testIdx}:${result.promptIdx}`;
 }
 
 function toResultMap(results: EvaluateResult[]): Map<string, EvaluateResult> {
